LayoutTests/fast/dom/HTMLInputElement/checkbox-dirty-checkedness-flag-expected.txt

Tests that setting the 'checked' content attribute does not set its dirty checkedness flag.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testCheckbox.checked is true
testCheckbox.removeAttribute('checked')
PASS testCheckbox.checked is false
testCheckbox.setAttribute('checked', '')
PASS testCheckbox.checked is true
PASS successfullyParsed is true

TEST COMPLETE
